requirements.txt
# Core FastAPI dependencies
fastapi[all]==0.116.2
uvicorn[standard]==0.35.0
pydantic==2.11.9
pydantic-settings==2.10.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20

# AWS Services
boto3==1.34.131
botocore==1.34.131

# Image Processing
Pillow==10.2.0

# Background Tasks
celery==5.3.4

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.2

# Monitoring & Logging
sentry-sdk[fastapi]==1.40.0

# Email
fastapi-mail==1.4.1

# HTTP Client
httpx==0.28.1

# Environment
python-dotenv==1.1.1

# Development tools
black==23.12.1
flake8==7.0.0
pre-commit==3.6.0